remove logs, fix css of filter, refactor
