+++ client/src/pages/ListView/ListView.css
@@ -86,109 +86,3 @@ html {
   border-top: 1px solid #dee2e6; /* Optional: Add a top border */
 }
 
-/* General styling for the search bar container */
-.search-bar-container {
-  display: flex;
-  align-items: center;
-  position: relative;
-  gap: 10px;
-}
-
-.search-bar {
-  width: 320px;
-  font-size: 15px;
-  border-radius: 8px;
-  border: 1px solid var(--color-secondary-300);
-  padding: 10px;
-  box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.1);
-  background-color: rgba(255, 255, 255, 0.397);
-}
-
-.filter-icon {
-  width: 25px;
-  height: 25px;
-  margin-left: 8px;
-  cursor: pointer;
-}
-
-/* Styling for the filter popup */
-.filter-popup {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  background: #fff;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  padding: 16px;
-  z-index: 10;
-  width: 320px;
-}
-
-.filter-fields {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.filter-field {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.filter-btn,
-.filter-btn:focus {
-  display: block;
-  margin: 4px;
-  width: 24px;
-  cursor: pointer;
-  border: none;
-  border-radius: 5px;
-  background-color: var(--color-primary-500);
-  color: var(--color-secondary-100);
-}
-
-.filter-btn:hover {
-  opacity: 0.8;
-}
-
-.filter-field label {
-  font-size: 14px;
-  margin-right: 8px;
-  flex-shrink: 0;
-  width: 80px;
-}
-
-.filter-field select,
-.filter-field input {
-  flex-grow: 1;
-  padding: 6px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-size: 14px;
-}
-
-.selected-filters {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-}
-
-
-/* Styling for applied filters under the search bar */
-.applied-filters {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 5px;
-  margin-top: 3px;
-}
-
-
-
-.filter-radio {
-  margin-bottom: 10px;
-  font-family: Arial, sans-serif;
-  font-size: 14px;
-  color: #333;
-}
